Improve local testing for all RDBMS (#75)
* Improve local testing for all RDBMS

* WIP

* WIP

* WIP

* WIP

* WIP

* Remove unused packages

* Remove comments

* WIP

* WIP

* Update Dockerfile


* Revert "Update Dockerfile"

This reverts commit ded8b1115f5437f36687743780bd00e9c1dde27e.

* Composer optimization

* WIP

* WIP

---------

+++ tests/Mysql/DatabaseTrait.php
@@ -17,7 +17,7 @@ trait DatabaseTrait
     {
         $database = getenv('YII_MYSQL_DATABASE');
         $host = getenv('YII_MYSQL_HOST');
-        $port = (int) getenv('YII_MYSQL_PORT');
+        $port = getenv('YII_MYSQL_PORT');
         $user = getenv('YII_MYSQL_USER');
         $password = getenv('YII_MYSQL_PASSWORD');
